branch-2.4: Merged everything in the trunk up to revision 1529.
git-svn-id: svn://localhost/public/MoSync/branches/release-2.4@1530 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ examples/QuakeMDL/inc/vector2i.hpp
@@ -118,7 +118,7 @@ public:
 
 	sint32 length ( void ) const
 	{
-		return (sint32)sqrt( this->dot( *this ) );
+		return (sint32)sqrt((double) this->dot( *this ) );
 	}
 
 	sint32 cross ( const Vector2i &v ) const
